io: move retry module to io/utils
