+++ io/itunes/fetchItunesItems.ts
@@ -3,7 +3,7 @@ import { createHash } from 'node:crypto'
 import transformCloudinaryImage from '@/io/cloudinary/transformCloudinaryImage'
 import { formatValidationError } from '@/utils/logging/zod'
 import { type Result, Ok, toErr } from '@/utils/errors/result'
-import { withRetry } from '@/io/retry'
+import { withRetry } from '@/io/utils/retry'
 import { filesystemCache, type CacheAdapter } from '@/io/cache/adapter'
 
 interface iTunesListItem {
