+++ io/notion/getBlockChildren.ts
@@ -8,7 +8,7 @@ import {
 } from './schemas/block'
 import { logValidationError } from '@/utils/logging/zod'
 import { Ok, toErr, type Result } from '@/utils/errors/result'
-import { withRetry } from '@/io/retry'
+import { withRetry } from '@/io/utils/retry'
 
 export const INVALID_BLOCK_ERROR = 'Invalid block data - build aborted'
