+++ io/notion/getMediaItems.ts
@@ -11,7 +11,7 @@ import { PageMetadataSchema } from './schemas/page'
 import { logValidationError } from '@/utils/logging/zod'
 import { env } from '@/io/env/env'
 import { type Result, Ok, toErr } from '@/utils/errors/result'
-import { withRetry } from '@/io/retry'
+import { withRetry } from '@/io/utils/retry'
 
 type MediaCategory = 'books' | 'albums' | 'podcasts'
