+++ io/tmdb/fetchTmdbList.ts
@@ -3,7 +3,7 @@ import transformCloudinaryImage from '@/io/cloudinary/transformCloudinaryImage'
 import { formatValidationError } from '@/utils/logging/zod'
 import { env } from '@/io/env/env'
 import { type Result, Ok, Err, toErr } from '@/utils/errors/result'
-import { withRetry } from '@/io/retry'
+import { withRetry } from '@/io/utils/retry'
 import { filesystemCache, type CacheAdapter } from '@/io/cache/adapter'
 
 // Schema for raw TMDB API response item
